+++ docs/changelog.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased
 
+***Added:***
+
+- Add `paths` parameter to the API
+
 ## 0.1.0 - 2024-09-22
 
 This is the initial public release.